requirements.txt
kornia==0.6.10
lpips==0.1.4
numpy==1.21.6
Pillow==12.2.0
scikit_image==0.19.3
scipy==1.7.3
segmentation_models_pytorch==0.3.2
tensorboardX==2.6.4
timm==0.6.12
torch==1.13.1+cu117
torchvision==0.14.1+cu117
